requirements.txt
# Meeting Transcriber - Dependencies
# Python 3.13+ compatible - minimal requirements without native dependencies

# OpenAI API (for Whisper transcription and GPT)
openai>=1.12.0

# Local Whisper transcription fallback (no API key required)
faster-whisper>=1.0.0

# GUI Framework
PySide6>=6.6.0

# Configuration management
python-dotenv>=1.0.0

# Audio processing (pure Python, no native dependencies)
numpy>=1.24.0
sounddevice>=0.4.6
soundcard>=0.4.2

# HTTP requests (already included with openai, but explicit for clarity)
requests>=2.31.0

# Note: Python's built-in 'wave' module is used for WAV file handling
# No external 'wave' package needed (the PyPI 'wave' package has bad dependencies)

# IMPORTANT: FFmpeg is REQUIRED for audio/video processing
# This is an external dependency (not a Python package)
# Install FFmpeg using:
#   Windows: winget install ffmpeg  OR  choco install ffmpeg
#   macOS:   brew install ffmpeg
#   Linux:   sudo apt install ffmpeg

# Optional: For advanced speaker diarization (requires GPU)
# pyannote.audio>=3.1.0
# torch>=2.0.0
